First commit of proposal for novel design.
Modular structure is proposed with a lightweight top level PRODUCT ontology. The ontology is aligned to both BOT and schema.org.

A taxonomy may be introduced as required.

This commit includes the initial commit of the P4Bldg ontology. A subset of the ifcOWL4Add2 ontology generated by the ProductOntologyGenerator.py script.